requirements.txt
feedparser==6.0.11
numpy==1.26.4
pyyaml==6.0.2
sentence-transformers==2.7.0
jinja2==3.1.4
python-dateutil==2.9.0
anthropic>=0.40.0
